skip existing-file check if cloned_item_id present
when replacing a file, we check whether the old content is in
use by any other Attachments before deleting it to avoid
orphaning it. but we can avoid this database query if the
attachment has a cloned_item_id; in that case, we know a clone
exists and the old content should not be deleted

this also ensures if a file somehow has a clone on another shard
sharing the same instfs_uuid (which shouldn't happen in ordinary
flow since #clone_for asks InstFS to duplicate the file in a
cross-shard clone) we don't accidentally that content

test plan:
 - have a file in a course
 - copy the course
 - overwrite the file with an identical one in the target
 - both copies should continue working
 - upload a new file to a course
 - re-upload the same file, overwriting it
 - the file's instfs_uuid should change and the service
   should be told to delete the old copy

flag=none
refs LF-1460

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/345146
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>

+++ app/models/attachment.rb
@@ -1803,6 +1803,7 @@ class Attachment < ActiveRecord::Base
   # after replacing this file's instfs_uuid, delete the old file if it's not being used by any other Attachments
   def safe_delete_overwritten_instfs_uuid(instfs_uuid)
     raise ArgumentError, "instfs_uuid not overwritten" if self.instfs_uuid == instfs_uuid
+    return if cloned_item_id.present?
 
     shard.activate do
       InstFS.delete_file(instfs_uuid) unless Attachment.where(instfs_uuid:).exists?
